[cli] Add a replay-events debug command.
This command renders a recorded stream of engine events using either the
progress or diff renderer. This is useful for debugging problems with
these renderers.

These changes also switch the JSON event log to show secrets. Without
rich secret information, the replay-events command cannot be
full-fidelity. Events sent to the service continue to be serialized
using the blinding encrypter, so all secret values are replaced with the
text "[secret]".